requirements.txt
torchvision
torch
Pillow
gdown
scikit-learn
torchviz